Multiple changes as suggested after review.
